convert to od_opinion
